pyproject.toml
[tool.poetry]
name = "relocation_helper"
version = "0.1.0"
description = ""
authors = ["cherkashin <alexander.cherkashin@sibit.tech>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
gevent = "^20.9.0"
bottle = "^0.12.19"
Paste = "^3.5.0"
dacite = "^1.6.0"
yoyo-migrations = "^7.2.1"
nltk = "^3.5"
pymystem3 = "^0.2.0"
systemd = "^0.16.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
